+++ cypress/e2e/awx/resources/inventories.cy.ts
