+++ cypress/e2e/awx/resources/projects.cy.ts
